fix for windows build.

+++ src/openflow/controllerApps/AbstractTCPControllerApp.cc
@@ -88,7 +88,7 @@ void AbstractTCPControllerApp::handleMessage(cMessage *msg){
                 event->setContextPointer(msg);
                 scheduleAt(simTime()+serviceTime, event);
             }
-            emit(queueSize,msgList.size());
+            emit(queueSize,static_cast<unsigned long>(msgList.size()));
             if(packetsPerSecond.count(floor(simTime().dbl())) <=0){
                 packetsPerSecond.insert(pair<int,int>(floor(simTime().dbl()),1));
             } else {
