+++ src/openflow/openflow/switch/engine/OF_Switch.cc
@@ -170,7 +170,7 @@ void OF_Switch::handleMessage(cMessage *msg){
                 event->setContextPointer(msg);
                 scheduleAt(simTime()+serviceTime, event);
             }
-            emit(queueSize,msgList.size());
+            emit(queueSize,static_cast<unsigned long>(msgList.size()));
             emit(bufferSize,buffer.size());
         }
     }
